+++ src/main/java/backend/models/Chemin.java
@@ -5,5 +5,5 @@ import java.util.*;
 public record Chemin(
 		List<Quai> cheminQuai,
 		List<Arete> cheminArete,
-		int poid // pour MeilleurCheminTemps, c'est le temps de trajet, pour MeilleurCheminCorrespondances, c'est le nombre de correspondances, pour MeilleurCheminCO2, c'est l'équivalent CO2 du trajet en grammes
+		double poid // pour MeilleurCheminTemps, c'est le temps de trajet, pour MeilleurCheminCorrespondances, c'est le nombre de correspondances, pour MeilleurCheminCO2, c'est l'équivalent CO2 du trajet en grammes
 		) {}
