+++ src/main/java/backend/models/DistanceAntecedants.java
@@ -5,24 +5,24 @@ import backend.models.Quai;
 import java.util.*;
 
 public class DistanceAntecedants {
-	private Integer distance;
+	private Double distance;
 	private List<Quai> antecedantsQuai;
 	private List<Arete> antecedantsArete;
 
-	public DistanceAntecedants(Integer distance, Quai antecedantQuai, Arete antecedantArete) {
+	public DistanceAntecedants(Double distance, Quai antecedantQuai, Arete antecedantArete) {
 		this.distance = distance;
 		this.antecedantsQuai = new ArrayList<>();
 		this.antecedantsArete = new ArrayList<>();
 		this.antecedantsQuai.add(antecedantQuai);
 		this.antecedantsArete.add(antecedantArete);
 	}
-	public DistanceAntecedants(Integer distance, List<Quai> antecedantsQuai, List<Quai> antecedantsArete) {
+	public DistanceAntecedants(Double distance, List<Quai> antecedantsQuai, List<Quai> antecedantsArete) {
 		this.distance = distance;
 		this.antecedantsQuai = new ArrayList(antecedantsQuai);
 		this.antecedantsArete = new ArrayList(antecedantsArete);
 	}
 
-	public Integer getDistance() {
+	public Double getDistance() {
 		return this.distance;
 	}
 	public List<Quai> getAntecedantsQuai() {
@@ -32,7 +32,7 @@ public class DistanceAntecedants {
 		return this.antecedantsArete;
 	}
 
-	public void setDistance(Integer distance) {
+	public void setDistance(Double distance) {
 		this.distance = distance;
 	}
 	public void setAntecedants(Quai antecedantQuai, Arete antecedantArete) {
